requirements.txt
# requirements.txt
flask
python-dotenv
pywin32
pyinstaller
tzdata